requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
